requirements.txt
aiogram==3.13.1
httpx==0.27.0
SQLAlchemy==2.0.32
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.8.2
backoff==2.2.1
uvloop==0.19.0
